+++ src/components/SubTaskList/SubTaskProgressBar/subTaskProgressBarUtils.js
